+++ webParent/WebBackEnd/src/main/java/com/onlineStore/admin/category/CategoryRestController.java
@@ -20,7 +20,7 @@ public class CategoryRestController {
     private static final Logger logger = LoggerFactory.getLogger(CategoryController.class);
 
     @PostMapping("/check_unique")
-    public String checkDuplicateEmail(@Param("id") Long id, @Param("name") String name, @Param("alies") String alies) {
+    public String checkDuplicateEmail(@Param("id") Integer id, @Param("name") String name, @Param("alies") String alies) {
 
         logger.info("Checking uniqueness for id={}, name={}, alies={}", id, name, alies);
